refractor(): split into modules
